requirements.txt
appdirs==1.4.4
Appium-Python-Client==2.11.1
async-generator==1.10
attrs==19.3.0
beautifulsoup4==4.11.1
black==22.1.0
bs4==0.0.1
castro==1.0.6
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.0.3
cryptography==3.4.8
distlib==0.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.0.12
h11==0.14.0
html-testRunner==1.2.1
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.4.0
mypy-extensions==0.4.3
names==0.3.0
openpyxl==3.0.10
outcome==1.3.0.post0
packaging==20.4
parameterized==0.7.4
pathspec==0.9.0
pi==0.1.2
platformdirs==2.4.1
pluggy==0.13.1
py==1.9.0
pycparser==2.21
pygame==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
pytz==2021.3
PyYAML==6.0
requests==2.31.0
selenium==4.11.2
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
timer==0.2.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.11.1
typed-ast==1.5.2
typing_extensions==4.0.1
urllib3==2.0.7
virtualenv==20.0.25
wcwidth==0.2.5
wsproto==1.2.0
zipp==3.1.0
